Fix problems with `Bound`
1. `Group.toBound`'s return type was wrongly marked `Group`. Closes #214
2. `Bound` stores top left and bottom right `Pt`s separately. Close #213

For 2, top left and bottom right `Pt`s were separate from `Pt`s that are
indexable using integers.

This causes that, when a calculation method such as `add` inherited from
`Group` is used on a `Bound`, the `Bound`'s corner `Pts`s are sometimes
not updated, as `Group` uses integer indices for getting the member
`Pt`s.

In order to fix this, the separate members `_topLeft` and `_bottomRight`
are removed from `Bound`. `topLeft` and `bottomRight` are now just
getters for integer-indexable `Pt`s.